consolidate citations;/
